add some type conversion

+++ SCI_Decode/src/Processor.cpp
@@ -207,7 +207,7 @@ void Processor::process_packet(SciFrame& frame, SciDataFile& datafile) {
                         cur_trigg_time_period_    = 0;
                         sci_trigger.time_period   = cur_trigg_time_period_;
                         sci_trigger.time_wait     = 0;
-                        cur_trigg_pre_raw_dead_   = sci_trigger.deadtime;
+                        cur_trigg_pre_raw_dead_   = static_cast<uint32_t>(sci_trigger.deadtime);
                         cur_trigg_num_g_ = 0;
                         sci_trigger.trigg_num_g   = cur_trigg_num_g_;
                         cur_trigg_num_g_++;
@@ -221,8 +221,8 @@ void Processor::process_packet(SciFrame& frame, SciDataFile& datafile) {
                         }
                         sci_trigger.time_period  = cur_trigg_time_period_;
                         sci_trigger.time_wait    = static_cast<uint32_t>(tmp_time_wait);
-                        int32_t tmp_dead_diff = static_cast<int32_t>(sci_trigger.deadtime) - cur_trigg_pre_raw_dead_;
-                        cur_trigg_pre_raw_dead_ = static_cast<int32_t>(sci_trigger.deadtime);
+                        uint32_t tmp_dead_diff = static_cast<uint32_t>(sci_trigger.deadtime) - cur_trigg_pre_raw_dead_;
+                        cur_trigg_pre_raw_dead_ = static_cast<uint32_t>(sci_trigger.deadtime);
                         tmp_dead_diff = (tmp_dead_diff > 0 ? tmp_dead_diff : tmp_dead_diff + 65536);
                         sci_trigger.dead_ratio = static_cast<float>(tmp_dead_diff) / static_cast<float>(sci_trigger.time_wait);
                         sci_trigger.trigg_num_g = cur_trigg_num_g_;
@@ -340,7 +340,7 @@ void Processor::process_packet(SciFrame& frame, SciDataFile& datafile) {
                         cur_event_time_period_[idx]     = 0;
                         sci_event.time_period           = cur_event_time_period_[idx];
                         sci_event.time_wait             = 0;
-                        cur_event_pre_raw_dead_[idx]    = sci_event.deadtime;
+                        cur_event_pre_raw_dead_[idx]    = static_cast<uint32_t>(sci_event.deadtime);
                         cur_event_num_g_[idx]           = 0;
                         sci_event.event_num_g           = cur_event_num_g_[idx];
                         cur_event_num_g_[idx]++;
@@ -354,8 +354,8 @@ void Processor::process_packet(SciFrame& frame, SciDataFile& datafile) {
                         }
                         sci_event.time_period  = cur_event_time_period_[idx];
                         sci_event.time_wait    = static_cast<uint32_t>(tmp_time_wait);
-                        int32_t tmp_dead_diff = static_cast<int32_t>(sci_event.deadtime) - cur_event_pre_raw_dead_[idx];
-                        cur_event_pre_raw_dead_[idx] = static_cast<int32_t>(sci_event.deadtime);
+                        uint32_t tmp_dead_diff = static_cast<uint32_t>(sci_event.deadtime) - cur_event_pre_raw_dead_[idx];
+                        cur_event_pre_raw_dead_[idx] = static_cast<uint32_t>(sci_event.deadtime);
                         tmp_dead_diff = (tmp_dead_diff > 0 ? tmp_dead_diff : tmp_dead_diff + 65536);
                         sci_event.dead_ratio = static_cast<float>(tmp_dead_diff) / static_cast<float>(sci_event.time_wait);
                         sci_event.event_num_g = cur_event_num_g_[idx];
